v2.4.16-First - See description
This update is one part of the 2.4.16 update. The second part will be released in 2 or 3 weeks.

Changelog :

New NMSWitherSkeletonBuilder class
New ConsumerCommand class
New setSilent method in the Utils class
Fixed /viapi credits message
Minor changes in some classes

Removed the deprecated percentageLuck method
Removed the deprecated PacketEntityHandler class

+++ src/net/vadamdev/viaapi/tools/math/MathF.java
@@ -67,7 +67,7 @@ public final class MathF {
     }
 
     public static float round(float a) {
-        return (float) Math.round(a);
+        return Math.round(a);
     }
 
     public static float random() {
